Agrega funcionalidades de criar e visualizar duvida, além do logout

+++ server/src/controler/UserControler.js
@@ -1,4 +1,3 @@
-//em todo lugar que era user coloquei account
 const account = require('../model/user')
 const jwt = require("jsonwebtoken");
 const secret = "mysecret";
@@ -17,23 +16,29 @@ module.exports = {
   },
 
   async createUser(req,res){ 
-      const { firstName, lastName, email, password, role } = req.body;
+      const { firstName, lastName, email, password, cPassword } = req.body;
 
       let generateData = {}
       generateData = {
-        firstName, lastName, email, password, role
+        firstName, lastName, email, password, cPassword
       }
 
-      //const usedEmail = await 
-      account.findOne({email:email}, function(err, account) {
-        if(account) {
-          res.status(403).json({message:"Email ja cadastrado." })
-        }
-        else {
-          const users = account.create(generateData); 
-          res.json(users) 
-        }
-      })
+      console.log(password, cPassword)
+
+      const usedEmail = await account.findOne({email})
+
+      if(usedEmail) {
+        res.status(403).json({message:"Email ja cadastrado." })
+      }
+      else if(password !== cPassword) {
+        res.status(403).json({message:"Senhas não conferem" })
+      }
+      else {
+        console.log(password, cPassword)
+        const users = account.create(generateData); 
+        res.json(users) 
+      }
+      
   },
 
   async update(req,res){
@@ -43,14 +48,20 @@ module.exports = {
     generateData = {
       firstName, lastName, email, password, role
     }
-    const users = await account.findByIdAndUpdate((_id), generateData, {new: true}); // SELECT * FROM users
+    if (_id.match(/^[0-9a-fA-F]{24}$/)) {
+      const users = await account.findByIdAndUpdate((_id), generateData, {new: true}); // SELECT * FROM users
     res.json(users)
+    }
+    
   },
 
   async delete(req,res){
     const {_id}= req.params
-    const users = await account.findByIdAndDelete({_id})// SELECT * FROM users where Id= ...
+    if (_id.match(/^[0-9a-fA-F]{24}$/)) {
+      const users = await account.findByIdAndDelete({_id})// SELECT * FROM users where Id= ...
     res.json(users)
+    }
+    
   },
 
   //lawyer controller 
@@ -81,8 +92,11 @@ module.exports = {
     generateData = {
       firstName, lastName, email, password, oabNumber, cpf, role
     }
-    const users = await account.findByIdAndUpdate((_id), generateData, {new: true}); // SELECT * FROM users
+    if (_id.match(/^[0-9a-fA-F]{24}$/)){
+      const users = await account.findByIdAndUpdate((_id), generateData, {new: true}); // SELECT * FROM users
     res.json(users)
+    }
+    
   },
 
   async login(req,res){
@@ -93,7 +107,7 @@ module.exports = {
         res.status(200).json({erro: "Erro no servidor, tente novamente"});
       }
       else if(!account) {
-        res.status(200).json({status:2, error: `${account}E-mail e/ou senha não conferem`})
+        res.status(200).json({status:2, error: `E-mail e/ou senha não conferem`})
       }
       else {
         account.isCorrectPassword(password, async function(err, same) {
@@ -104,7 +118,7 @@ module.exports = {
             res.status(200).json({status: 2, error:"SENHAE-mail e/ou senha não conferem."})
           }
           else {
-            const payload = {email};
+            const payload = {_id};
             token = jwt.sign(payload, secret, {
               expiresIn: '24h'
             });
@@ -128,6 +142,38 @@ module.exports = {
             }
         })
     }
+  },
+  async teste(req,res){
+    const user = await account.findById(req.user._id)
+
+  if (user) {
+    res.json({
+      _id: user._id,
+      name: user.firstName,
+      email: user.email,
+    })
+  } else {
+    res.status(404)
+    throw new Error('User not found')
+  }
+    /*const token = req.body.token || req.query.token || req.cookies.token || req.headers['x-access-token'];
+    if(!token){
+        //res.json({status:401,msg:'Não autorizado: Token inexistente!'});
+        try{
+          const decoded = jwt.verify(token, secret)
+          console.log(decoded)
+          next();
+        }
+        catch(err){
+          console.log(err)
+        }
+        
+                res.json({status:401,msg:'Não autorizado: Token inválido!'});
+            }else{
+                res.json({status:200})
+            }
+        })
+    }*/
   },
   async logout(req,res){
       const token = req.headers.token;
